requirements.txt
python-binance
RPLCD
RPi.GPIO